Rewrite to use adaptive characteristics #15
Where we earlier had hardcoded characteristics, we now search for the
uuids in dbus and return the corresponding paths.
This is done on device creation, and we also check that the paths are
nonzero when we connect. It seems like these occasionaly dissapear from
the the bluetooth adapter, when it hasnt been connected for some time.