Dont create multiple timeouts. Fixes #20
The old code would create multiple check timeout tasks, which would all
run in parallell and subtract from the same timeout value,
leading to this value reaching 0 much quicker than intended.

Change this to only create a new timeout task if none exists, or the
previous one has completed.

Shrink the timeout to 3 sec to be more snappy, and add a timeout update to messages
sent, not just received