Fix reconnect logic, add stability #24
This commit does a lot, but I was not able to separate it in a
meaningful way.

Important parts:

Add reconnect logic where there earlier was simply an infinite
wait/sleep loop. Make sure there are only onle notify listener active at
a time

Add a signal handler which listens to changed to the Connected property from dbus instead of setting it
to true and false ourselves

Filter which tasks we cancel when disconnecting, so we dont cancel
requests we receive while disconnecting. The window is about 3 seconds,
so pretty significant, and since we want the app to not hold the
connection too long, this can happen when a user has finished
inputing a pin

Add try/except blocks to calls that go to dbus

Finally, add timestamp to logging