Add pairing updates to app. Closes #11
Adds a monitor that tracks the interfaces added and removed to the bluez
dbus objectmanager.
Filter the added interfaces for fido devices, and setup another
listener that listens to changes to the "Paired" property.
The listener callback then updates the list of devices.

We make the fido_devices and hid_devices global, so the callbacks can
access them without passing arguments everywhere.

The interfaces removed callback checks if the removed interface
belonged to a device we had stored, and then delete it from our lists.

The new update_fido_devices method checks the devices we get from
the objectmanager against those stored in the app to avoid duplicates

Finally, we set the self.connected flag to true at the end of the
connect method. At this point we have gathered all characteristics from
the device, so we know we are connected.
We need to set this here, because our 'Connected' listener only listens to changes in the property.
However, after we pair we are already connected, so this won't happen until we disconnect.
Without explicitly setting the flag here, we need to disconnect after
pairing to use the device.